move func defs up another level
