+++ src/CommonFunctions.jl
@@ -1,0 +1,43 @@
+function add_click_event end
+function add_collision_event end
+function append_array end
+function append_array end
+function apply_forces end
+function center_text end
+function check_collisions end
+function destroy end
+function draw end
+function flip end
+function get_last_update end
+function get_offset end
+function get_parent end
+function get_position end
+function get_rotation end
+function get_scale end
+function get_size end
+function get_tag end
+function get_type end
+function get_velocity end
+function handle_event end
+function initialize end
+function load_image end
+function load_sound end
+function render end
+function set_color end
+function set_last_update end
+function set_offset end
+function set_parent end
+function set_position end
+function set_rotation end
+function set_scale end
+function set_size end
+function set_sprite end
+function set_tag end
+function set_vector2_value end
+function set_vector2f_value end
+function stop_music end
+function toggle_sound end
+function unload_sound end
+function update end
+function update_array_value end
+function update_text end
