+++ src/Component/ComponentFunctions.jl
@@ -1,39 +1,39 @@
 # Declare Common Functions so that they can be dispatched from ModuleExtensions
-function add_collision_event end
-function append_array end
-function append_array end
-function apply_forces end
-function check_collisions end
-function destroy end
-function draw end
-function draw end
-function flip end
-function get_last_update end
-function get_offset end
-function get_parent end
-function get_position end
-function get_rotation end
-function get_scale end
-function get_size end
-function get_tag end
-function get_type end
-function get_velocity end
-function initialize end
-function load_image end
-function load_sound end
-function set_color end
-function set_last_update end
-function set_offset end
-function set_parent end
-function set_position end
-function set_rotation end
-function set_scale end
-function set_size end
-function set_sprite end
-function set_tag end
-function set_vector2f_value end
-function stop_music end
-function toggle_sound end
-function unload_sound end
-function update end
-function update_array_value end
+import ..JulGame: add_collision_event,
+    append_array,
+    append_array,
+    apply_forces,
+    check_collisions,
+    destroy,
+    draw,
+    draw,
+    flip,
+    get_last_update,
+    get_offset,
+    get_parent,
+    get_position,
+    get_rotation,
+    get_scale,
+    get_size,
+    get_tag,
+    get_type,
+    get_velocity,
+    initialize,
+    load_image,
+    load_sound,
+    set_color,
+    set_last_update,
+    set_offset,
+    set_parent,
+    set_position,
+    set_rotation,
+    set_scale,
+    set_size,
+    set_sprite,
+    set_tag,
+    set_vector2f_value,
+    stop_music,
+    toggle_sound,
+    unload_sound,
+    update,
+    update_array_value
